The progress line can now be composed.
The user can extend the progress line without changing the library
itself. She has to implement the `ProgressLinePart` interface and create
a progress line with her parts.

    new ProgressLine(new MyFirstPart(), new MySecondPart()).

The existing `Counter` is currently the only part that is provided by
JFortschritt. (The progress bar is not a part, because it is the heart
of the progress line and cannot be skipped.)